Register: Add input validation code
Ensure that all the entries have been filled when 'register' is pressed.
If that is not the case, then do not actually register, but display an
error and indicate which fields still need to be filled

Created an Admin class, which currently is no different than User, but
will be in the future. Updated the switch(..) in Manager to handle this
type and construct it