Cleanup: Fix minor bugs, cleanup code
Fixed bug with title for the titlebar in the ShelterActivity code. The
setSupportActionBar(...) needs to be called after setting the title
text.

Converted the initialization loop within ShelterActivity into a cleaner
lambda expression, as well as, inverting the if condition to reduce
nesting.

Set the TextView in ShelterActivity to the empty string rather than just
displaying "TextView" above the Android logo.

Added a back functionality for the ListActivity class to go back to the
main application page.