Search: Fix reassignment bug
Fixed bug where shelterList was reassigned to point to a new list;
however, the RecyclerView still has the reference to the old list.

Instead, shelterList is cleared and repopulated with all the available
shelters, and then filtered based on what was typed into the search box.

Therefore the reference stays constant, but the elements in the list
that it points to are being changed, making the changes visible to the
RecyclerView when it it notified.

(cherry picked from commit 8ce153f05dd11a77e2098e68d4b602b6d70a3c94)