DB: Overhaul previous database implementation
The Database abstract class (previously: interface) was completely
redesigned, adding features like key sets, value lists, and removing
pairings.

CsvDB is a key-value DB that is sourced from a .csv file. It contains
the required CSV parsing code and is powered by an Encoder which takes
the parsed Strings and formulates an Object to be stored within the
in-memory k-v store.

ListActivity was modified to contain a name-based search bar as well as
a gender and age-range spinner, which is a more intuitive UI than the
previously used "omnibar" approach.

The name-based search is very specific: contains() is not employed,
since that allows an out-of-order match. Instead, a substring starting
from the *beginning* of the string is created and directly matched
against the inputted text.