Search: Add Gender and Age-based search options
Added gender and age-based search options, via spinners, to the
ListActivity. The 'Any' options for both gender and age are currently
working: it's effectively a no-op with the name-based search doing all
the work.

The Toolkit class contains helpful methods, like union(...) and
intersect(...) which perform the set operations they are named after,
which makes multi-option search easy to implement.