Login: Add Firebase authentication support
Added support for persistant account creation and login using the
Firebase Authentication. This replaces the previous setup, which was
neither persistant nor full functional and flexible.

The main WelcomeActivity has been stripped down and automatically pulls
up the AuthUI fragment. The Login and Register activities still exist,
but will be removed in the near future.