Shelter: Move shelter data to Firebase
Moved the shelter data from a local .csv into the Firebase real-time
database. The application loads the shelter data from the database and
displays it in a RecyclerView, like before.

The capacity field in the CSV has been made into a new Capacity class
which contains the capacity "type" and the associated amount (int) of
that type. Displaying this information still needs to be done; however,
retrieving that information from Firebase is already functional.