Layout: Display capacity information
Added support for displaying the capacity information. The capacity
along with the type are displayed. The ability to display multiple
capacities for a single shelter has also been implemented.

If there is no known capacity (unspecified in the original .csv data),
then "Unknown (?)" is displayed to the user. The user would most likely
have to call the shelter (using the provided number) to get
clarification and reserve space, if required.