Reserve: Implement reservation system
Implemented the ability to reserve beds and rooms for the different
shelters. Successful reservation will update Firebase and the
application will immediately see the changes.

Reservation is based on types: e.g. beds, apartments, rooms, etc., so
there is an individual capacity and available field for each type that
the particular shelter offers.